# frontend/src/shared/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "wat2do API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/ai/generate-filters": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Generate Filters",
        "operationId": "generate_filters_ai_generate_filters_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIPromptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilterStateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/ai/generate-event": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Generate Event",
        "operationId": "generate_event_ai_generate_event_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIPromptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventFormDataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/ai/parse-event-image": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Parse Event Image",
        "operationId": "parse_event_image_ai_parse_event_image_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_parse_event_image_ai_parse_event_image_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventFormDataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/auth/send-otp": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Send Otp",
        "operationId": "send_otp_auth_send_otp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendOtpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/verify-otp": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Verify Otp",
        "operationId": "verify_otp_auth_verify_otp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "description": "Log the user out.\n\nA11: Bearer is optional \u2014 if the access token has expired, the refresh\ncookie is still cleared so the browser doesn't hold a replayable token.\nEven if upstream revocation fails, the cookie is still cleared; we\nsurface the error via a JSONResponse that carries Set-Cookie for\n``refresh_token``.",
        "operationId": "logout_auth_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/calendar/token": {
      "get": {
        "tags": [
          "calendar"
        ],
        "summary": "Get Calendar Token",
        "description": "Return the user's feed token, generating one on first call.",
        "operationId": "get_calendar_token_calendar_token_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarTokenResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/calendar/token/regenerate": {
      "post": {
        "tags": [
          "calendar"
        ],
        "summary": "Regenerate Calendar Token",
        "description": "Rotate the user's feed token; existing subscriptions break.",
        "operationId": "regenerate_calendar_token_calendar_token_regenerate_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarTokenResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/calendar/feed/{token}.ics": {
      "get": {
        "tags": [
          "calendar"
        ],
        "summary": "Get Calendar Feed",
        "description": "Public: return the VCALENDAR for the token's owner.\n\nCalendar clients (Google, Apple, Outlook) poll this URL and render\neach VEVENT as a calendar entry.  Rate-limited per token.",
        "operationId": "get_calendar_feed_calendar_feed__token__ics_get",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AIPromptRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Prompt"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "AIPromptRequest"
      },
...
      "Body_parse_event_image_ai_parse_event_image_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_parse_event_image_ai_parse_event_image_post"
      },
...
      "CalendarTokenResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "feed_url": {
            "type": "string",
            "title": "Feed Url"
          }
        },
        "type": "object",
        "required": [
          "token",
          "feed_url"
        ],
        "title": "CalendarTokenResponse",
        "description": "Returned by GET /calendar/token and POST /calendar/token/regenerate."
      },
...
      "EventFormDataResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "occurrences": {
            "items": {
              "$ref": "#/components/schemas/EventFormOccurrenceResponse"
            },
            "type": "array",
            "title": "Occurrences"
          },
          "location": {
            "type": "string",
            "title": "Location",
            "default": ""
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": ""
          },
          "price": {
            "type": "number",
            "title": "Price",
            "default": 0
          },
          "food": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Food"
          },
          "registration": {
            "type": "boolean",
            "title": "Registration",
            "default": false
          },
          "source_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Image Url"
          }
        },
        "type": "object",
        "title": "EventFormDataResponse"
      },
      "EventFormOccurrenceResponse": {
        "properties": {
          "dtstart_local": {
            "type": "string",
            "title": "Dtstart Local",
            "default": ""
          },
          "dtend_local": {
            "type": "string",
            "title": "Dtend Local",
            "default": ""
          }
        },
        "type": "object",
        "title": "EventFormOccurrenceResponse"
      },
...
      "FilterStateResponse": {
        "properties": {
          "searchQuery": {
            "type": "string",
            "title": "Searchquery",
            "default": ""
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Categories",
            "default": []
          },
          "locations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Locations",
            "default": []
          },
          "foods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Foods",
            "default": []
          },
          "days": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Days",
            "default": []
          },
          "priceRange": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Pricerange",
            "default": {
              "min": "",
              "max": ""
            }
          },
          "dateRange": {
            "type": "string",
            "title": "Daterange",
            "default": ""
          },
          "addedSince": {
            "type": "string",
            "title": "Addedsince",
            "default": ""
          },
          "registration": {
            "type": "boolean",
            "title": "Registration",
            "default": false
          }
        },
        "type": "object",
        "title": "FilterStateResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse"
      },
...
      "SendOtpRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token",
            "description": "Optional invitation token"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "SendOtpRequest"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "school": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "School"
          },
          "onboarding_required": {
            "type": "boolean",
            "title": "Onboarding Required",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_in",
          "user_id"
        ],
        "title": "TokenResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifyOtpRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "token": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "email",
          "token"
        ],
        "title": "VerifyOtpRequest"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}